MCP/pyproject.toml
[project]
name = "mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=1.46.0",
    "python-dotenv>=1.2.1",
]
